split remote watcher stuff out of main.
the main.rs file was defining a whole bunch of remote tracking logic. it
shouldn't need to have that responsibility
