+++ src/caseta/mod.rs
@@ -1,7 +1,9 @@
 mod connection;
 mod message;
 mod scene;
+mod remote;
 
 pub use connection::*;
 pub use message::*;
 pub use scene::*;
+pub use remote::*;
